After missing over 6hrs

+++ EmployeeManagement.Web/Pages/DisplayEmployeeBase.cs
@@ -29,12 +29,19 @@ namespace EmployeeManagement.Web.Pages
             await OnEmployeeSelection.InvokeAsync((bool)e.Value);
         }
 
-
-        protected async Task Delete_Click()
+        protected MagnusQ.Components.ConfirmBase DeleteConfirmation { get; set; }
+        protected void Delete_Click()
         {
-            await EmployeeService.DeleteEmployee(Employee.EmployeeId);
-            await OnEmployeeDeleted.InvokeAsync(Employee.EmployeeId);
-            //NavigationManager.NavigateTo("/", true);
+            DeleteConfirmation.Show();
+        }
+
+        protected async Task ConfirmDelete_Click(bool deleteConfirmed)
+        {
+            if(deleteConfirmed)
+            {
+                await EmployeeService.DeleteEmployee(Employee.EmployeeId);
+                await OnEmployeeDeleted.InvokeAsync(Employee.EmployeeId);
+            }
         }
     }
 }
